refactor: exrtact request with timeout and retry as a stateless function
